fix(interpret): break circular import by moving interpret.ts to api-indexer/
Root cause: `core/index.ts` re-exported `core/interpret.ts`, which at module
load time imported `BitBadgesCollection` as a runtime value from
`api-indexer/`. BitBadgesCollection's transitive deps cycle back into
`core/`, so by the time its class body evaluated, `CollectionDoc` was
still undefined — producing `TypeError: Super expression must either be
null or a function` at load.

Symptom: 37 test suites in bitbadges-indexer were failing identically
with this error, blocking CI entirely once `test.yml`'s exit code was
fixed to actually gate on results (see bitbadges-indexer PR #102).

Fix: interpret.ts operates on a BitBadgesCollection instance, so it
structurally belongs next to the class, not in `core/`. Moved:
  - src/core/interpret.ts       -> src/api-indexer/interpret.ts
  - src/core/interpret.spec.ts  -> src/api-indexer/interpret.spec.ts

Relative imports inside the moved file adjusted accordingly.
Re-exports: `core/index.ts` no longer exports `./interpret.js`;
`api-indexer/index.ts` now does. Top-level `import { interpretCollection }
from 'bitbadges'` is unchanged for consumers. Helpers re-exported by
interpret.ts (timestampToDate, durationToHuman, denomToHuman,
amountToHuman) remain accessible from the top-level export.

Also updated three direct importers that had hard-coded paths:
  - src/builder/tools/builders/explainCollection.ts
  - src/cli/commands/builder.ts
  - src/cli/commands/sdk.ts

Verification:
- `bun run build` passes (tsc + tsc-alias + fixup + dep-check all green).
- `madge --circular` confirms: "No circular dependency found!"
- Manual CJS load test: `typeof BitBadgesCollection === 'function'`
  (was throwing "Super expression" before).
